fix: preserve fork title across metadata updates
